Score System--Hellfire
Added a score system to add score on defeating enemies and collecting coins.

The player drops the number of coins assigned in the Score System from the inspector on death. Currently, death is when the shield gets to 0.

Added score to the Save/Load System.

Level to test - Scenes/Testing/Hellfire_ScoreSystem