Shield and Shield Break  - Hellfire
Enemy attacks player on coming in contact with them and applies knockback force.

The force increases with every hit and is capped to a max value.

If the enemy hits the player after the shield is cracked (shield health = 0), the player dies and it's Game Over.

Level to test - Scenes/Feature Implementation/Hellfire/Shield and Shield Break.